#39 adds support for more LEDsections in CLOCK mode
This basicly multiplies the LEDs,
For example when you have a section of 3 it means that LED 1+2+3 now behave as one section, and that the clock will treat them as such. Resulting in a clock with tripple the amount of LEDs in this case.

Default is stil 1 section, so 60 LEDs in total, this is just support for if you want to make a hardcoded bigger clock